refactor : #18 CreateMemberRequest 클래스 구현
